GetProfile

+++ androidApp/src/main/java/com/example/socialmediaapp/android/account/follows/Following.kt
@@ -11,7 +11,7 @@ import org.koin.androidx.compose.koinViewModel
 @Destination<RootGraph>
 fun Following(
     navigator: DestinationsNavigator,
-    userId: Int
+    userId: Long
 ) {
     val viewModel: FollowsViewModel = koinViewModel()
