+++ androidApp/src/main/java/com/example/socialmediaapp/android/account/follows/FollowsViewModel.kt
@@ -16,7 +16,7 @@ class FollowsViewModel : ViewModel(){
     var uiState by mutableStateOf(FollowsUiState())
         private set
 
-    fun fetchFollows(userId: Int, followsType: Int){
+    fun fetchFollows(userId: Long, followsType: Int){
         viewModelScope.launch {
             uiState = uiState.copy(isLoading = true)
